Clean up test duplication
Test CLI had a subset of already existing tests found in test_server

Looking at the tests, they are all very much server based tests, so
consolidated into test_server and removed test_cli

We should look into increasing more cli tests
